refactor: Move the HostFunctionArguments class
Move it into the BinaryInjection.Loader namespace
