+++ src/CoreHook.BinaryInjection/Loader/HostFunctionArguments.cs
@@ -1,6 +1,7 @@
 ﻿using System.IO;
+using CoreHook.BinaryInjection.Loader.Serializer;
 
-namespace CoreHook.BinaryInjection.Loader.Serializer
+namespace CoreHook.BinaryInjection.Loader
 {
     public class HostFunctionArguments : IBinarySerializer
     {
