+++ src/CoreHook.BinaryInjection/RemoteInjection/RemoteInjector.cs
@@ -6,7 +6,6 @@ using System.Reflection;
 using System.Runtime.InteropServices;
 using CoreHook.BinaryInjection.Loader;
 using CoreHook.BinaryInjection.Loader.Serializer;
-using CoreHook.BinaryInjection.Host;
 using CoreHook.CoreLoad.Data;
 using CoreHook.IPC.Platform;
 using CoreHook.Memory;
